LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/query-container-name-expected.txt


FAIL match closest named container assert_equals: expected "yes" but got "no"
FAIL match ancestor named container assert_equals: expected "yes" but got "no"
PASS no match for unused container name
